+++ snowflake/ml/jobs/_utils/scripts/constants.py
@@ -2,3 +2,9 @@
 SHUTDOWN_ACTOR_NAME = "ShutdownSignal"
 SHUTDOWN_ACTOR_NAMESPACE = "default"
 SHUTDOWN_RPC_TIMEOUT_SECONDS = 5.0
+
+
+# Log start and end messages
+# Inherited from snowflake.ml.jobs._utils.constants
+LOG_START_MSG = "--------------------------------\nML job started\n--------------------------------"
+LOG_END_MSG = "--------------------------------\nML job finished\n--------------------------------"