+++ snowflake/ml/model/_client/sql/service.py
@@ -1,5 +1,4 @@
 import enum
-import json
 import textwrap
 from typing import Any, Optional, Union
 
@@ -15,22 +14,25 @@ from snowflake.snowpark import dataframe, functions as F, row, types as spt
 from snowflake.snowpark._internal import utils as snowpark_utils
 
 
+# The enum comes from https://docs.snowflake.com/en/sql-reference/sql/show-service-containers-in-service#output
+# except UNKNOWN
 class ServiceStatus(enum.Enum):
     UNKNOWN = "UNKNOWN"  # status is unknown because we have not received enough data from K8s yet.
     PENDING = "PENDING"  # resource set is being created, can't be used yet
-    READY = "READY"  # resource set has been deployed.
     SUSPENDING = "SUSPENDING"  # the service is set to suspended but the resource set is still in deleting state
     SUSPENDED = "SUSPENDED"  # the service is suspended and the resource set is deleted
     DELETING = "DELETING"  # resource set is being deleted
     FAILED = "FAILED"  # resource set has failed and cannot be used anymore
     DONE = "DONE"  # resource set has finished running
-    NOT_FOUND = "NOT_FOUND"  # not found or deleted
     INTERNAL_ERROR = "INTERNAL_ERROR"  # there was an internal service error.
+    RUNNING = "RUNNING"
+    DELETED = "DELETED"
 
 
 class ServiceSQLClient(_base._BaseSQLClient):
     MODEL_INFERENCE_SERVICE_ENDPOINT_NAME_COL_NAME = "name"
     MODEL_INFERENCE_SERVICE_ENDPOINT_INGRESS_URL_COL_NAME = "ingress_url"
+    SERVICE_STATUS = "service_status"
 
     def build_model_container(
         self,
@@ -199,22 +201,17 @@ class ServiceSQLClient(_base._BaseSQLClient):
         include_message: bool = False,
         statement_params: Optional[dict[str, Any]] = None,
     ) -> tuple[ServiceStatus, Optional[str]]:
-        system_func = "SYSTEM$GET_SERVICE_STATUS"
-        rows = (
-            query_result_checker.SqlResultValidator(
-                self._session,
-                f"CALL {system_func}('{self.fully_qualified_object_name(database_name, schema_name, service_name)}')",
-                statement_params=statement_params,
-            )
-            .has_dimensions(expected_rows=1, expected_cols=1)
-            .validate()
-        )
-        metadata = json.loads(rows[0][system_func])[0]
-        if metadata and metadata["status"]:
-            service_status = ServiceStatus(metadata["status"])
-            message = metadata["message"] if include_message else None
-            return service_status, message
-        return ServiceStatus.UNKNOWN, None
+        fully_qualified_object_name = self.fully_qualified_object_name(database_name, schema_name, service_name)
+        query = f"SHOW SERVICE CONTAINERS IN SERVICE {fully_qualified_object_name}"
+        rows = self._session.sql(query).collect(statement_params=statement_params)
+        if len(rows) == 0:
+            return ServiceStatus.UNKNOWN, None
+        row = rows[0]
+        service_status = row[ServiceSQLClient.SERVICE_STATUS]
+        message = row["message"] if include_message else None
+        if not isinstance(service_status, ServiceStatus):
+            return ServiceStatus.UNKNOWN, message
+        return ServiceStatus(service_status), message
 
     def drop_service(
         self,