+++ tests/integ/snowflake/ml/jobs/test_constants.py
@@ -1,6 +1,6 @@
 from snowflake.ml._internal.utils import snowflake_env
 
-_TEST_COMPUTE_POOL = "E2E_TEST_POOL"
+_TEST_COMPUTE_POOL = "E2E_CPU_POOL"
 _TEST_SCHEMA = "ML_JOB_TEST_SCHEMA"
 _SUPPORTED_CLOUDS = {
     snowflake_env.SnowflakeCloudType.AWS,